update:异步测试方法
